+++ src/common/components/index.js
@@ -5,3 +5,4 @@ export * from './FieldClaimingDate';
 export * from './LineLocationsView';
 export * from './NumberGeneratorModal';
 export * from './NumberGeneratorButton';
+export * from './RemoveFromPackageModals';