+++ src/common/hooks/index.js
@@ -11,6 +11,8 @@ export * from './usePiecesExpect';
 export * from './usePiecesRequests';
 export * from './useReceive';
 export * from './useReceivingTenantIdsAndLocations';
+export * from './useRemoveFromPackage';
 export * from './useTitle';
 export * from './useTitleHydratedPieces';
+export * from './useTitleMutation';
 export * from './useUnreceive';